Address comments in review, 2; reduce differences between container and executable build of PyPerf

+++ scripts/pyperf_build.sh
@@ -5,6 +5,21 @@
 #
 set -euo pipefail
 
+if [ "$#" -gt 1 ]; then
+    echo "Too many arguments"
+    exit 1
+elif [ "$#" -eq 0 ]; then
+    with_staticx=""
+elif [ "$1" == "--with-staticx" ]; then
+    with_staticx="$1"
+else
+    echo "Unexpected argument: $1"
+    exit 1
+fi
+
+if [ "$(uname -m)" = "aarch64" ]; then
+    exit 0;
+fi
 git clone --depth 1 -b upgrade-bcc-reference https://github.com/marcin-ol/bcc.git && cd bcc && git reset --hard 863d545
 
 # (after clone, because we copy the licenses)
@@ -19,3 +34,13 @@ mkdir build
 cd build
 cmake -DPYTHON_CMD=python3 -DINSTALL_CPP_EXAMPLES=y -DCMAKE_INSTALL_PREFIX=/bcc/root -DBCC_CLOADER_KERNEL_HEADERLESS=1 ..
 make -C examples/cpp/pyperf -j -l VERBOSE=1 install
+# leave build directory
+cd ..
+# leave bcc repository
+cd ..
+
+# We're using staticx to build a distribution-independent binary of PyPerf because PyPerf
+# can only build with latest llvm (>10), which cannot be obtained on CentOS.
+if [ -n "$with_staticx" ]; then
+    staticx ./root/share/bcc/examples/cpp/PyPerf ./root/share/bcc/examples/cpp/PyPerf
+fi
